Properly add last lap data
Polar does not auto add last lap to the list, so app has to create it
on it's own. Right now the code is still a little bit messy. The output
directory is hardcoded, and the ConvertCommand does a little bit too
much logic. Will work on that in the next commits.

Right now the data gathered by app can be successfuly added to Endomondo
and looks good.